Fixed offset by 1 error in SimplePaste. Removed warnings
git-svn-id: https://subversion.assembla.com/svn/ATD_Libraries/FreeImageAlgorithms@785 422d053a-e635-4566-91ac-90b073d95016

+++ trunk/src/FreeImageAlgorithms_Arithmetic.cpp
@@ -364,12 +364,12 @@ template < class Tsrc > int ARITHMATIC < Tsrc >::AddImages (FIBITMAP * dst, FIBI
     }
     else
     {
-        float *dst_ptr;
+        double *dst_ptr;
 
         for(register int y = 0; y < height; y++)
         {
 
-            dst_ptr = (float *) FreeImage_GetScanLine (dst, y);
+            dst_ptr = (double *) FreeImage_GetScanLine (dst, y);
             src_ptr = (Tsrc *) FreeImage_GetScanLine (src, y);
 
             for(register int x = 0; x < width; x++)
@@ -424,12 +424,12 @@ template < class Tsrc > int ARITHMATIC < Tsrc >::SubtractImages (FIBITMAP * dst,
     }
     else
     {
-        float *dst_ptr;
+        double *dst_ptr;
 
         for(register int y = 0; y < height; y++)
         {
 
-            dst_ptr = (float *) FreeImage_GetScanLine (dst, y);
+            dst_ptr = (double *) FreeImage_GetScanLine (dst, y);
             src_ptr = (Tsrc *) FreeImage_GetScanLine (src, y);
 
             for(register int x = 0; x < width; x++)
@@ -580,12 +580,12 @@ template < class Tsrc > int ARITHMATIC < Tsrc >::AddGreyLevelImageConstant (FIBI
     }
     else
     {
-        float *dst_ptr;
+        double *dst_ptr;
 
         for(register int y = 0; y < height; y++)
         {
 
-            dst_ptr = (float *) FreeImage_GetScanLine (dst, y);
+            dst_ptr = (double *) FreeImage_GetScanLine (dst, y);
 
             for(register int x = 0; x < width; x++)
                 dst_ptr[x] += casted_constant;
@@ -633,11 +633,11 @@ template < class Tsrc > int ARITHMATIC < Tsrc >::SubtractGreyLevelImageConstant 
     }
     else
     {
-        float *dst_ptr;
+        double *dst_ptr;
 
         for(register int y = 0; y < height; y++)
         {
-            dst_ptr = (float *) FreeImage_GetScanLine (dst, y);
+            dst_ptr = (double *) FreeImage_GetScanLine (dst, y);
 
             for(register int x = 0; x < width; x++)
                 dst_ptr[x] -= casted_constant;
